requirements.txt
greenlet==3.2.3
pillow==11.3.0
playwright==1.53.0
pyee==13.0.0
python-dotenv==1.1.1
typing_extensions==4.14.1
requests==2.32.3
